Don't source argcomplete if bash < 4

+++ zendev/bootstrap.sh
@@ -11,7 +11,9 @@ zendev () {
 
 alias cdz="zendev cd"
 
-ZENDEV_ARGCOMPLETE=$(mktemp -u /tmp/zd.argcomplete.XXXXX)
-activate-global-python-argcomplete --dest=- > "${ZENDEV_ARGCOMPLETE}"
-source "${ZENDEV_ARGCOMPLETE}"
-rm -f "${ZENDEV_ARGCOMPLETE}"
+if [ ${BASH_VERSION:0:1} -ge 4 ]; then
+    ZENDEV_ARGCOMPLETE=$(mktemp -u /tmp/zd.argcomplete.XXXXX)
+    activate-global-python-argcomplete --dest=- > "${ZENDEV_ARGCOMPLETE}"
+    source "${ZENDEV_ARGCOMPLETE}"
+    rm -f "${ZENDEV_ARGCOMPLETE}"
+fi
